major file refactor, clared contained, moved class to utils, merged parsers
